requirements.txt
grpcio
googleapis-common-protos
qrcode
flask
gunicorn
